use DEPENDENCIES in bootstrapping

+++ .Rprofile
@@ -6,10 +6,17 @@ if (!"https://rse.pik-potsdam.de/r/packages" %in% getOption("repos")) {
 
 # bootstrapping, will only run once after this repo is freshly cloned
 if (isTRUE(rownames(installed.packages(priority = "NA")) == "renv")) {
-  message("R package dependencies are not installed in this renv, installing now...")
-  renv::install("yaml", prompt = FALSE) # yaml is required to find dependencies in Rmd files
-  renv::hydrate() # auto-detect and install all dependencies
-  message("Finished installing R package dependencies.")
+  if (file.exists("DESCRIPTION") && tools::md5sum("DEPENDENCIES") != tools::md5sum("DESCRIPTION")) {
+    warning("Unexpected DESCRIPTION file found, try removing it. Will not install dependencies.")
+  } else {
+    message("R package dependencies are not installed in this renv, installing now...")
+    renv::install("yaml", prompt = FALSE) # yaml is required to find dependencies in Rmd files
+    # renv will only consider a file called DESCRIPTION, so rename DEPENDENCIES
+    file.copy("DEPENDENCIES", "DESCRIPTION")
+    renv::hydrate() # auto-detect and install all dependencies
+    file.remove("DESCRIPTION")
+    message("Finished installing R package dependencies.")
+  }
 }
 
 # source global .Rprofile (very important to load user specific settings)
